the_requirements.txt
absl-py==1.2.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.5
attr==0.3.2
backcall==0.2.0
black==21.4b2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
decorator==5.1.1
detectron2==0.5+cu110
einops==0.4.1
executing==0.9.1
fonttools==4.34.4
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.48.0
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iopath==0.1.8
ipython==8.4.0
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mmcv-full==1.3.0
# Editable Git install with no remote (mmsegmentation==0.11.0)
-e /cluster/home/guosun/azc/vss/VSS-CFFM
MultiScaleDeformableAttention==1.0
mypy-extensions==0.4.3
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.5.1
prompt-toolkit==3.0.30
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
six==1.16.0
stack-data==0.3.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
timm==0.6.7
toml==0.10.2
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
unfoldNd==0.2.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.1
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
